+++ VendingMachine/src/ServiceLayer/VendingMachineServiceLayerInterface.java
@@ -14,7 +14,7 @@ import java.util.ArrayList;
  * @author nacer
  */
 public interface VendingMachineServiceLayerInterface {
-   public Change computeChange(float userMoney, float itemCost) throws InsufficientFundsException;
+   public Change computeChange(double userMoney, double itemCost) throws InsufficientFundsException;
    public void updateItem(String itemName) throws NoItemInventoryException, CannotOpenFile; // Updates an item, we will just remove one from the stock
    public Item getSpecificItem(String itemName); // Gets a specific item using it's name from the ArrayList
    public ArrayList<Item> displayableItemDetailList(); // Returns a list of all displayable items
